First attempt with pulling from my own repo:
Noticed its a little hard to setup a new build of the OS,
so i added some scripts. under tools, gencrt will pull the
i686-elf-gcc's crtbegin.o and crtend.o files and place them
in the ix86 directory. This is so the linker can link those
in the end, since git decides to gitignore .o files i guess.
in scripts folder, genbindir will generate the full structure
to the bin directory, so that the folders can be matched and
files can be compiled into their respected binary folders.
Added a whole cpufeatures source file because the featureString
function was acting weird.


Signed-off-by: climjark <climjark@protonmail.com>